float doble fix

+++ trunk/po0_220222/task_03/src/Circle.cpp
@@ -20,12 +20,12 @@ void Circle::SetRadius(const float _radius)
 
 void Circle::CountArea()
 {
-	SetArea((pow(GetRadius(), 2) * std::numbers::pi));
+	SetArea(float((pow(GetRadius(), 2) * std::numbers::pi)));
 }
 
 void Circle::CountPerimeter()
 {
-	SetPerimeter(2 * std::numbers::pi * GetRadius());
+	SetPerimeter(float(2.00 * std::numbers::pi * GetRadius()));
 }
 
 void Circle::Print() const
